break out opportunity base from admin
